+++ sample-app/src/test/java/com/fabernovel/statefullayout/utils/clicks.kt
@@ -1,2 +1,9 @@
 package com.fabernovel.statefullayout.utils
 
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.ViewInteraction
+import androidx.test.espresso.action.ViewActions
+import androidx.test.espresso.matcher.ViewMatchers
+
+fun clickOn(resId: Int): ViewInteraction = onView(ViewMatchers.withId(resId))
+    .perform(ViewActions.click())
